refactor: [#109] subclassing으로 중복 코드 정리
